Mise à jour des types d'identifiants de usize à i32 dans les structures et ajout de la fonction get_tuto pour récupérer un tutoriel par son identifiant.

+++ model/src/model.rs
@@ -4,26 +4,26 @@ pub mod model {
 
     #[derive(Debug, Serialize, Deserialize)]
     pub struct Id {
-        pub value: usize,
+        pub value: i32,
     }
 
     #[derive(Debug, Serialize, Deserialize)]
     pub struct Tuto {
-        pub id: usize,
+        pub id: i32,
         pub title: String,
         pub content: String,
     }
 
     #[derive(Debug, Serialize, Deserialize)]
     pub struct Tag {
-        pub tuto_id: usize,
+        pub tuto_id: i32,
         pub value: String,
     }
 
     #[derive(Debug, Serialize, Deserialize)]
     pub struct Resultat {
         pub score: i8, // nombre de tags trouvés pour ce tuto
-        pub tuto_id: usize,
+        pub tuto_id: i32,
         pub tags: Vec<String>, // tags trouvés parmi les arguments
         pub title: String,
         pub content: String,
